+++ server/src/services/signatureHelp.ts
@@ -2,7 +2,7 @@ import {SymbolFunction} from "../compiler_analyzer/symbolObject";
 import {Position, SignatureHelp, URI} from "vscode-languageserver";
 import {findScopeContainingPosition} from "./serviceHelper";
 import {ParameterInformation, SignatureInformation} from "vscode-languageserver-types";
-import {ComplementKind, ComplementCallerArgument} from "../compiler_analyzer/symbolComplement";
+import {ComplementKind, ComplementCallerArgument} from "../compiler_analyzer/complementHint";
 import {resolveTemplateType, stringifyResolvedType} from "../compiler_analyzer/symbolUtils";
 import {SymbolScope} from "../compiler_analyzer/symbolScope";
 import {TextPosition} from "../compiler_tokenizer/textLocation";
